+++ src/hpack/huffman_constants.py
@@ -1,8 +1,4 @@
-# -*- coding: utf-8 -*-
 """
-hpack/huffman_constants
-~~~~~~~~~~~~~~~~~~~~~~~
-
 Defines the constant Huffman table. This takes up an upsetting amount of space,
 but c'est la vie.
 """