+++ src/hpack/struct.py
@@ -1,13 +1,15 @@
-# -*- coding: utf-8 -*-
 """
-hpack/struct
-~~~~~~~~~~~~
-
 Contains structures for representing header fields with associated metadata.
 """
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    from typing_extensions import Self, TypeAlias  # pragma: no cover
 
 
-class HeaderTuple(tuple):
+class HeaderTuple(tuple[bytes, bytes]):
     """
     A data structure that stores a single header field.
 
@@ -21,11 +23,12 @@ class HeaderTuple(tuple):
     This class stores a header that can be added to the compression context. In
     all other ways it behaves exactly like a tuple.
     """
+
     __slots__ = ()
 
     indexable = True
 
-    def __new__(cls, *args):
+    def __new__(cls, *args: Any) -> Self:
         return tuple.__new__(cls, args)
 
 
@@ -34,6 +37,14 @@ class NeverIndexedHeaderTuple(HeaderTuple):
     A data structure that stores a single header field that cannot be added to
     a HTTP/2 header compression context.
     """
+
     __slots__ = ()
 
     indexable = False
+
+    def __new__(cls, *args: Any) -> Self:
+        return tuple.__new__(cls, args)
+
+
+Header: TypeAlias = "HeaderTuple | NeverIndexedHeaderTuple | tuple[bytes, bytes]"
+HeaderWeaklyTyped: TypeAlias = "HeaderTuple | NeverIndexedHeaderTuple | tuple[bytes | str, bytes | str]"