+++ tests/test_hpack.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import itertools
 import pytest
 
@@ -16,6 +15,14 @@ from hpack import (
     InvalidTableSizeError,
 )
 from hpack.hpack import _dict_to_iterable, _to_bytes
+
+
+def test_to_bytes():
+    assert _to_bytes(b"foobar") == b"foobar"
+    assert _to_bytes("foobar") == b"foobar"
+    assert _to_bytes(0xABADBABE) == b'2880289470'
+    assert _to_bytes(True) == b'True'
+    assert _to_bytes(Encoder()).startswith(b"<hpack.hpack.Encoder")
 
 
 class TestHPACKEncoder: