+++ CHANGELOG.rst
@@ -6,11 +6,21 @@ dev
 
 **API Changes (Backward Incompatible)**
 
--
+- Support for Python 3.6 has been removed.
+- Support for Python 3.7 has been removed.
+- Support for Python 3.8 has been removed.
+- Renamed `InvalidTableIndex` exception to `InvalidTableIndexError`.
 
-**Bugfixes**
+**API Changes (Backward Compatible)**
 
--
+- Support for Python 3.9 has been added.
+- Support for Python 3.10 has been added.
+- Support for Python 3.11 has been added.
+- Support for Python 3.12 has been added.
+- Support for Python 3.13 has been added.
+- Optimization of headers into bytes encoding.
+- Added type hints.
+- Updated packaging and testing infrastructure.
 
 
 4.0.0 (2020-08-30)